% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlpls.R
\name{rlpls}
\alias{rlpls}
\title{A main function of RLPLS}
\usage{
rlpls(
  X,
  Y,
  maxk = 5,
  eta0 = 0.4,
  kppa0 = 0.3,
  Btimes = 1000,
  b_var_num = c(NA, NA),
  var.selection_alpha = 0.05,
  model_fit = "simpls",
  variable_select = "simpls",
  scale.x = TRUE,
  scale.y = TRUE,
  maxiter = 5000,
  eps0 = 1e-04,
  seed = 2022
)
}
\arguments{
\item{X}{Matrix of predictors.}

\item{Y}{Vector or matrix of responses.}

\item{maxk}{The number of hidden components, with default 5.}

\item{eta0}{A thresholding parameter between 0 and 1, with default 0.4.}

\item{kppa0}{A parameter between 0 and 0.5 managing the effect of the concavity of the objective function and the closeness of direction vectors and surrogate vectors. It affects a model for multivariate response variables. The default is 0.3.}

\item{Btimes}{The number of boostrapping for the first direction vector.}

\item{b_var_num}{A vector indicating the number of selected variables for bootstrapping. The length should be 2 and each element needs to be smaller than the number of predictor variables.}

\item{var.selection_alpha}{A significance level for variable selection.}

\item{model_fit}{A PLS algorithm for fitting a model. The alternatives are \code{"kernelpls"}, \code{"widekernelpls"}, \code{"simpls"}, or \code{"oscorespls"}. The default is \code{"simpls"}. The selection is independent from the selection of \code{variable_select}.}

\item{variable_select}{A PLS algorithm for variable selection. The alternatives are \code{"pls2"} or \code{"simpls"}. The default is \code{"simpls"}. The selection is independent from the selection of \code{model_fit}.}

\item{scale.x}{logical value whether to scale predictor variables.}

\item{scale.y}{logical value whether to scale response variables.}

\item{maxiter}{The maximum number of iterations when fitting direction vectors, with default is 5000.}

\item{eps0}{An effective zero, with default 1e-4.}

\item{seed}{A seed number.}
}
\value{
a list of rlpls result.
  \item{beta.rlpls}{ The estimated coefficients from RLPLS.}
  \item{dataset}{ The data used for the model.}
  \item{pls.fit.method}{ A fitting method used for the model.}
  \item{variable.select.method}{ A variable selection method used for the model.}
}
\description{
This is a main function of sparse partial least square regression with random LASSO (RLPLS).
}
\examples{
set.seed(111)
# Set the size of samples and true coefficients
n = 30
beta0 = c(runif(20,-10,-5),runif(20,5,10),rep(0,10))

# Construct predictors & response
# the structure is similar with that from Chun and Keles (2010)
xxx = yyy = c(NULL)
H1 = c(rep(2,times=25),rep(4,times=n-25))
H2 = rep(10,n)
for(i in 1:length(beta0)){
  if(i <= 30){
    xxx = cbind(xxx, H1+rnorm(n,0,1))
  }else{
    xxx = cbind(xxx, H2+rnorm(n,0,1))
  }
};rm(i)# for i

yyy = xxx \%*\% beta0 + rnorm(n,0,1.5^2)

# Run RLPLS
library(rlpls)
result = rlpls(X = xxx, Y = yyy, b_var_num = c(25,20), Btimes = 5000 , variable_select = "pls2")
result$beta.rlpls

}
